Move mapper -> pkg/mapper
to make clear that this is meant to be used elsewhere
