+++ pkg/mapper/telemetry.go
